requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
packaging==23.0
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyuploadcare==4.0.0
replicate==0.5.2
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tailwindcss==0.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
